apps/api/requirements.txt
fastapi==0.115.5
uvicorn[standard]==0.32.1
pydantic[email]==2.10.3
pydantic-settings==2.6.1
python-dotenv==1.0.1

# ─── Database ───
sqlalchemy[asyncio]==2.0.36
asyncpg==0.30.0
alembic==1.14.0
greenlet==3.1.1

# ─── Auth ───
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-multipart==0.0.17

# ─── Dev / tests ───
pytest==8.3.4
pytest-asyncio==0.24.0
httpx==0.28.1
